feat(reservation): 예약 관련 기능 추가 및 개선
- 예약 생성 및 취소 로직 구현
- 예약 상태 변경 시 이벤트 발행 로직 추가
- Redis를 이용한 예약 정보 저장 및 조회 기능 구현
- 예약 관련 DTO 및 인터페이스 추가

+++ src/main/java/depth/finvibe/investment/modules/asset/application/AssetEventService.java
@@ -1,5 +1,6 @@
 package depth.finvibe.investment.modules.asset.application;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 import jakarta.transaction.Transactional;
@@ -56,7 +57,7 @@ public class AssetEventService implements AssetEventUseCase {
         return PortfolioGroupDto.RegisterAssetRequest.builder()
                 .stockId(event.getStockId())
                 .name(event.getName())
-                .stockPrice(event.getPrice())
+                .stockPrice(BigDecimal.valueOf(event.getPrice()))
                 .amount(event.getAmount())
                 .currency(Currency.valueOf(event.getCurrency()))
                 .build();
@@ -65,7 +66,7 @@ public class AssetEventService implements AssetEventUseCase {
     private PortfolioGroupDto.UnregisterAssetRequest createUnregisterRequestFrom(TradeExecutedEvent event) {
         return PortfolioGroupDto.UnregisterAssetRequest.builder()
                 .stockId(event.getStockId())
-                .stockPrice(event.getPrice())
+                .stockPrice(BigDecimal.valueOf(event.getPrice()))
                 .amount(event.getAmount())
                 .currency(Currency.valueOf(event.getCurrency()))
                 .build();
